Move RelocationModifier back to linker_utils
I need it there so that I can use it from linker-diff.
