+++ libwild/src/arch.rs
@@ -7,6 +7,7 @@ use crate::error::Result;
 use crate::resolution::ValueFlags;
 use anyhow::bail;
 use linker_utils::elf::SectionFlags;
+use linker_utils::relaxation::RelocationModifier;
 use std::borrow::Cow;
 use std::str::FromStr;
 
@@ -67,10 +68,4 @@ pub(crate) trait Relaxation {
     fn debug_kind(&self) -> impl std::fmt::Debug;
 
     fn next_modifier(&self) -> RelocationModifier;
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum RelocationModifier {
-    Normal,
-    SkipNextRelocation,
 }
