+++ libwild/src/elf_writer.rs
@@ -7,7 +7,6 @@ use self::elf::TLS_MODULE_BASE_SYMBOL_NAME;
 use crate::alignment;
 use crate::arch::Arch;
 use crate::arch::Relaxation as _;
-use crate::arch::RelocationModifier;
 use crate::args::Args;
 use crate::args::BuildIdOption;
 use crate::args::FileWriteMode;
@@ -74,6 +73,7 @@ use linker_utils::elf::shf;
 use linker_utils::elf::sht;
 use linker_utils::elf::RelocationKind;
 use linker_utils::elf::SectionFlags;
+use linker_utils::relaxation::RelocationModifier;
 use memmap2::MmapOptions;
 use object::elf::NT_GNU_BUILD_ID;
 use object::elf::NT_GNU_PROPERTY_TYPE_0;
