+++ libwild/src/layout.rs
@@ -11,7 +11,6 @@ use crate::alignment;
 use crate::alignment::Alignment;
 use crate::arch::Arch;
 use crate::arch::Relaxation as _;
-use crate::arch::RelocationModifier;
 use crate::args::Args;
 use crate::args::BuildIdOption;
 use crate::args::OutputKind;
@@ -73,6 +72,7 @@ use itertools::Itertools;
 use linker_utils::elf::shf;
 use linker_utils::elf::RelocationKind;
 use linker_utils::elf::SectionFlags;
+use linker_utils::relaxation::RelocationModifier;
 use object::elf::gnu_hash;
 use object::elf::Rela64;
 use object::elf::GNU_PROPERTY_AARCH64_FEATURE_1_AND;
