+++ linker-utils/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod elf;
+pub mod relaxation;
 pub mod x86_64;
