+++ linker-utils/src/relaxation.rs
@@ -1,0 +1,5 @@
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum RelocationModifier {
+    Normal,
+    SkipNextRelocation,
+}
